Move classes to new namespace
For some reason, I stuck them at the top level of o.l.aikuma, which was
silly. Fixed now.
